Beta quality now.
XMI fast loading support now uses a Sax Parser so that we don't have to
deal with encoding. 
Provided a builder API to configure which content-types or file
extension to crawl.

I did test performances on a large workspace with thousands of models.
The XMI-fast mode takes a few seconds to initialize and subsequent calls
are instantaneous (yeah !)

Using content-types makes things slower during the initialization (file
extensions are faster to check !)